New license command
This commit adds a license command that fetches license information from
the mgmtd and prints the contained information as a human readable
overview or as json. Optionally, license information reload on the
server can be triggered by adding the `--reload` flag.

It uses grpc for mgmtd communication and depends on licensing and
certificate related types in the protobuf repository.