Remove globalConfig and use viper for backend configuration
Viper already provides a global config store, so why not use it. This
provides the benefits of environment variable bindings (and potential
config file bindings) actually working and viper taking care of all of
that.

A major advantage is also that values can be checked for being set by
the user or using the default value. This allows for easier optional
values and is, for example, important for cases where a given file does
not exist is an error but not if it's the default file.

To avoid mistyping the strings I defined constants for the valid viper
keys which are currently also used to provide the flag names (that is
not a must have, but makes the flag definitions less verbose).