Rework/export mappers and add mappings type
Originally mapper types were only part of the entries package and not designed to be exported. For example they used global singleton instances that had to be carefully initialized before first use. As they are used in more places this would become increasingly error prone and difficult to debug.

Now instead of using global singleton instances, callers are responsible for initializing and passing around instances of individual mappers or using the combined mappings type if multiple mappers are required.

General improvements:

- Functions to initialize mappers no longer need to make network requests to obtain the details needed to initialize each mapper. Instead this is provided by the caller, optimizing use cases like GetMappings that can collect these details once and reuse them to initialize multiple mappers.
- Internal mapper maps are no longer exported and instead only accessed using public methods. For example there are new methods that allow iterating over each of the entity ID types using the for range syntax.
- Common MapX functions no longer unnecessarily loop over each of the entity ID types and instead create the mappings all at once.

New mappers:

- Mapping targets to their entity ID set.
- Mapping nodes to their targets.

Fixes:

- Previously mapping meta buddy groups to their primary nodes did not actually map to the entity ID set of the node, but rather the primary target. While the legacy ID was correct by coincidence, the alias and UID were not.