Remove old index update functionality and alias update to create (#92)
The old update mode would start the same process as the update service that runs constantly to listen for modification events.
This is probably not what most users would expect, they are probably wanting to trigger a manual re-index of an existing BeeGFS.
Drop the old update functionality from CTL and alias the update command to create which is the correct way to manually re-index.