Fix BadgerDB logs ignoring logging configuration
Because BadgerDB logs are very spammy (especially at debug level) they were hardcoded to INFO.
This inadvertently overrode the custom logger causing Badger to revert back to the default logging behavior (log to stderr).
Note it is also not possible to set the log level before the custom logger as the setting will be ignored.

The long-term fix will be to allow logging verbosity for different components to be set independently of the overall log level (https://github.com/ThinkParQ/beegfs-go/issues/109).