fix: improve license output in case of errors (#141)
There are multiple classes of errors that can occur during license
reload and verification and that we need to handle and report to the user
as clearly as possible.

The specific issue this commit addresses is that in case of a failed
reload, the new certificate data wasn't printed at all, which suggests
to the user that the mgmtd didn't in fact load the new certificate.
That is one possibility, but if the loading succeeds (license file and
library exist, file is readable and parses as x509) and the only issue
is that the certificate is not valid, the management will actually keep
the newly loaded certificate regardless of verification status.

To reflect that to the user, we will now try to load the data again if
we encounter an error during reload and display the new certificate data
if available. If no data can be loaded, an appropriate error will be
printed to inform the user about that.