fix(ctl): return a question mark when the chunk path cannot be calculated
When querying verbose entry info, in some scenarios the timestamp could be invalid which would cause a panic:

  panic: runtime error: slice bounds out of range [:-4]

This would happen when querying entry info for a dangling dentry which is possible when querying entry info while deleting the same files.

Errors generating verbose output should not stop execution but instead be logged. This was not actually happening due to a separate bug where the wrong error was logged.