fix(ctl): strengthen path handling
(1) Reject fs provider init when the user cannot access the specified path. When initializing a fs
provider from a user-specified path, if the path is non-existent or temporarily inaccessible the
logic walks up the tree to attempt finding a valid path inside a BeeGFS mount point. This change
ensures initialization is rejected if access to the user specified path fails due to permission
errors. This ensures calling code doesn't accidentally assume access a user shouldn't have.

(2) Add a euid check in the backend when initializing the global mount results in an unmounted file
system. This ensures only users with root permissions can access unmounted file systems, even if CTL
is used as a library and the caller does not check this.