feat(ctl): add the ability to GetEntryInfo using an ioctl
With this change users without root privileges can no longer use an RPC to fetch entry info. This
strengthens path handling to ensure users cannot infer the existence of files in BeeGFS that they do
not have permission to access, even by guessing paths or filenames.

On most system the ioctl path is expected to be more performant due to the effects of client
metadata caching. However testing shows on some systems the RPC path is still faster so a new option
was added to skip using an ioctl and use direct RPCs even if BeeGFS is mounted. The RPC path is
likely only faster on systems where the ctl+client+servers are on the same machine (i.e., test
systems) so there is virtually no network overhead and the syscall overhead is greater.