feat(rst): add GetStubContents rpc to remote
The GetStubContents rpc allows clients to request from remote the rstId and
remote path contained in a stub for an offloaded file. This is necessary
since non-hsm client mounts will not have read access to the stub file.

Update remote job requests to utilize GetStubContents for offloaded files
instead of configured targets.