feat(ctl): make `stats client|user` compatible with ipv6 server nodes (#229)
The IPv6 capable servers can report IPv6 addresses, which is only possible with a new
GetClientStatsV2 message. To maintain compatibility with not updated legacy nodes, this includes a
mechanism to detect whether or not use the new V2 message or the old one, so this should work with a
mixed system of pre- and post IPv6 nodes.