fix(rst): avoid silently dropping errors building job requests
When an entry had no RST IDs configured there was no way to communicate back when an error occurred
getting locked info. The fix differentiates between fatal errors getting locked info (which should
always terminate the builder job) and errors that are specific to a particular entry that should be
reported on a per-entry basis when possible.

Without this fix, if calling GetEntry returned an error and there were no user specified RSTs,
builder jobs could fail with a vague: 'job builder failed to complete: no matches found in local
path: /downloads; --remote-target was not provided so relying on configured rstIds and stub urls'.

Note at this time the specific sequence of events that leads to the vague error are pretty unlikely
outside contrived tests, as GetEntry failing likely means the file system is inaccessible and the
job would have failed much earlier in a more obvious manner.